extract linux/macos readers
